+++ Assembly-CSharp/UI/InGame/PlayerVoteArea.cs
@@ -244,11 +244,11 @@ public class PlayerVoteArea : MonoBehaviour
 				PlayerSkin = CreateNewSprite(PlayerIcon);
 
 
-				PlayerControl.SetHatImage(hatID, PlayerHat, 0, playerControl.Data.ColorId);
-				PlayerControl.SetHatImage(hatID, PlayerHatExt, 1, playerControl.Data.ColorId);
-				PlayerControl.SetHatImage(hatID, PlayerHatExt2, 2, playerControl.Data.ColorId);
-				PlayerControl.SetHatImage(hatID, PlayerHatExt3, 3, playerControl.Data.ColorId);
-				PlayerControl.SetHatImage(hatID, PlayerHatExt4, 4, playerControl.Data.ColorId);
+				PlayerControl.SetHatImage(hatID, PlayerHat, 0, (int)playerControl.Data.ColorId);
+				PlayerControl.SetHatImage(hatID, PlayerHatExt, 1, (int)playerControl.Data.ColorId);
+				PlayerControl.SetHatImage(hatID, PlayerHatExt2, 2, (int)playerControl.Data.ColorId);
+				PlayerControl.SetHatImage(hatID, PlayerHatExt3, 3, (int)playerControl.Data.ColorId);
+				PlayerControl.SetHatImage(hatID, PlayerHatExt4, 4, (int)playerControl.Data.ColorId);
 				PlayerControl.SetSkinImage(playerControl.Data.SkinId, PlayerSkin);
 				IconsLoaded = true;
 			}
